Moved TagsMatcher.
